Rename file to PluggableSearchBarControls. Remove use memo
